added baasic heap functiobnality(not implemeneted) and moved agloritmhs out of main
